BackEnd/requirements.txt
fastapi
pydantic
pymongo
python-dotenv
uvicorn
pyjwt
passlib[bcrypt]
datetime
cryptography
openai
dotenv
pdf2image
python-multipart